Prepare the models for multi-year support
Add a year feld to all relevant models so that
we can assign Devrooms, Exhibitors and Sponsorings
as well as Sponsoring packages to the year they belong
to.

Additionally add a new ConferenceYear model to the main app
that will be used to store all the years that are available
for listing.