Remove old site-local config example
Due to refactoring the old site-local config is now obsolete
the new config is in local_settings.py.example. The settings.py
is now static for all installations and should not be touched
anymore.

Moreover, add the local_settings.py to the example symlink to the
ignorelist, so others can remove this when needed for their
development environment.